chore: port reset password to features
