+++ apps/admin/src/app/resetPassword/layout.tsx
@@ -4,13 +4,6 @@ import React, { useEffect } from "react";
 import ThemeSwitcher from "@/components/theme-switcher";
 
 const LoginLayout = ({ children }) => {
-  useEffect(() => {
-    if (typeof localStorage !== "undefined") {
-      if (localStorage.theme === "dark") {
-        ThemeSwitcher.switch("dark");
-      }
-    }
-  }, []);
   return (
     <>
       <div className="">{children}</div>
