+++ apps/admin/src/features/post/index.ts
@@ -1,0 +1,2 @@
+export { Editor } from "./components/editor";
+export { Feature } from "./feature";
