+++ apps/admin/src/features/post/views/default.tsx
@@ -5,7 +5,7 @@ import { FontPageWrapper } from "@/components/fonts";
 
 import { Post, PostStatusOptions, Setting } from "@/__generated__/__types__";
 
-import Editor from "../components/editor";
+import { Editor } from "../components/editor";
 import { FeaturedImage } from "../components/featured-image";
 import { SubTitle } from "../components/subtitle";
 import { Title } from "../components/title";
