+++ apps/admin/src/features/reset-password/index.ts
@@ -1,0 +1,1 @@
+export { Feature } from "./feature";
